Added tests for AuthNetAccountController

+++ app/controllers/admin/AuthNetAccountController.java
@@ -10,9 +10,11 @@ import com.fasterxml.jackson.databind.JsonNode;
 import be.objectify.deadbolt.java.actions.Group;
 import be.objectify.deadbolt.java.actions.Restrict;
 import errors.ValidationError;
+import forms.CreditCardForm;
 import models.AuthNetAccount;
 import models.json.MessageResponse;
 import models.json.ObjectCreatedResponse;
+import play.Logger;
 import play.data.DynamicForm;
 import play.data.FormFactory;
 import play.libs.Json;
@@ -43,7 +45,15 @@ public class AuthNetAccountController extends Controller {
 	public Result addAuthNetAccount() {
 		
 		JsonNode json = request().body().asJson();
-		AuthNetAccount account = Json.fromJson(json, AuthNetAccount.class);
+		
+		AuthNetAccount account = null;
+		try {
+			account = Json.fromJson(json, AuthNetAccount.class);
+		} 
+		catch (RuntimeException e) {
+			Logger.error("Cannot parse JSON to Merchant Account.");
+			return badRequest(Json.toJson(new MessageResponse("ERROR", "Cannot parse JSON to Merchant Account")));
+		}
 		
 		List<ValidationError> errors = authNetAccountService.validate(account);
     	if (errors != null) {
@@ -61,7 +71,15 @@ public class AuthNetAccountController extends Controller {
 	@BodyParser.Of(BodyParser.Json.class)
 	public Result editAuthNetAccount() {
 		JsonNode json = request().body().asJson();
-		AuthNetAccount account = Json.fromJson(json, AuthNetAccount.class);
+		
+		AuthNetAccount account = null;
+		try {
+			account = Json.fromJson(json, AuthNetAccount.class);
+		} 
+		catch (RuntimeException e) {
+			Logger.error("Cannot parse JSON to Merchant Account.");
+			return badRequest(Json.toJson(new MessageResponse("ERROR", "Cannot parse JSON to Merchant Account")));
+		}
 		
 		List<ValidationError> errors = authNetAccountService.validate(account);
     	if (errors != null) {
@@ -69,20 +87,34 @@ public class AuthNetAccountController extends Controller {
     		return badRequest(Json.toJson(errors));
     	}
 		
-    	authNetAccountService.update(account);
+    	AuthNetAccount accountDB = authNetAccountService.getById(account.getId());
+    	if (accountDB == null) {
+	    	return badRequest(Json.toJson(new MessageResponse("ERROR", "Merchant Account with id " + 
+	    			account.getId() + " is not found")));
+	    }
+    	
+    	authNetAccountService.update(accountDB);
 		
 		return ok(Json.toJson(new MessageResponse("SUCCESS", "Merchant Account was edited successfully")));
 	}
+	
 	public Result deleteAuthNetAccount() {
 		
 		DynamicForm form = formFactory.form().bindFromRequest();
 		long id = Long.parseLong(form.get("id"));
 		
 		AuthNetAccount account = authNetAccountService.getById(id);
-		authNetAccountService.delete(account);
+		if (account == null) {
+	    	return badRequest(Json.toJson(new MessageResponse("ERROR", "Merchant Account with id " + 
+	    			id + " is not found")));
+	    }
 		
-		return ok(Json.toJson(new MessageResponse("SUCCESS", "Merchant Account was deleted successfully")));
-		
+		boolean deleted = authNetAccountService.delete(account);
+		if (deleted) {
+			return ok(Json.toJson(new MessageResponse("SUCCESS", "Merchant Account was deleted successfully")));
+		} else {
+			return badRequest(Json.toJson(new MessageResponse("ERROR", "Error occured while deleting the Merchant Account")));
+		}		
 	}
 	
 }
